requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pydantic[email]

pyjwt
bcrypt
pycrypto

python-dotenv
pytest
